Showcase pattern matching bug
