+++ tests/tests/bugfixes.rs
@@ -24,3 +24,31 @@ fn test_pattern_binding_bug() {
 
     assert_eq!(out, 3 * 11);
 }
+
+/// Bug where string patterns just do not work when part of a match statement
+/// inside of an inst fn call.
+#[test]
+fn test_string_pattern_in_instance_fn_bug() {
+    rune! {
+        () =>
+        enum Inst { A, Unknown }
+
+        pub fn works() {
+            let program = [];
+            let inst = match "a" { "a" => Inst::A, _ => Inst::Unknown };
+            program.push(inst);
+            assert_eq!(program, [Inst::A]);
+        }
+
+        pub fn broken() {
+            let program = [];
+            program.push(match "a" { "a" => Inst::A, _ => Inst::Unknown });
+            assert_eq!(program, [Inst::A]);
+        }
+
+        pub fn main() {
+            works();
+            broken();
+        }        
+    };
+}
